Formalize expect callback response with `%Moxinet.Response{}` (#20)
Background:
It became appearant that the simple map, returned from an `expect/4`
callback
wasn't enough to make an interface that was easy to grasp and limit
potential mistakes.

This change will introduce `%Moxinet.Response{}`, which will be the
expected return from all
callbacks sent to `Moxinet.expect/4`. With it, we'll get some help from
the compiled by getting
compilation errors when trying to pass invalid fields and therefore make
it easier to work with.