Allow `times` option to `expect` to control usage limits (#29)
Background:
In scenarios where you'd want to test a retry pattern or have mocks that
responds differently each time you run them you'd need to be able to
have multiple mocks defined for each signature.

One of the core concepts of multiple mocks is to define how many times a
mock can be reused, which historically have been only once. Subsequent
calls to the same mock would result in an error.

This change will allow a mock to be used multiple times through the
`times` options on the `expect` function.

To achieve a reasonable API, the other optional arguments (`pid` and
`storage_pid`) have also been moved to `options`, meaning we'd see a
breaking API change.