Add `Moxinet.ReqTestAdapter` to integrate fully into Req (#65)
Background:
The `req` library, which is one of the most popular ones out there
today, have this nifty ability of defining an custom adaptor, which will
make it possible to customize the request and response steps of Req.

This means we can leverage this to automatically set the moxinet header
and also capture special responses from moxinet, which can be turned
into more suitable errors on the client side

Changes:
- Allow signature storage to be configured on mocks (mainly allows
better integration testing inside moxinet)
- Add the `Moxinet.ReqTestAdapter` which both sets the moxinet pid
header and captures + reraises errors delivered from moxinet.