Raise error when expectations in a test were not fully used (#22)
Background:
Unused expectations can happen through copy-pasting or when code
does unexpected things. This means having some help with removing
unused expectations, or verifying that the ones that are defined, are
also used, will increase the overall health of a testsuite.

This change adds the `verify_usage!` function, which can be used to
verify that all expectations were ran, and defines an `ExUnit` `on_exit`
callback
to automatically run it